+++ libft/ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: emadriga <emadriga@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/28 17:03:35 by emadriga          #+#    #+#             */
-/*   Updated: 2021/06/04 18:45:33 by emadriga         ###   ########.fr       */
+/*   Updated: 2021/11/09 11:46:10 by emadriga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,14 @@
 
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	while ((*s1 != '\0' || *s2 != '\0') && n-- > 0)
-		if (*s1++ != *s2++)
-			return ((unsigned char)*(s1 - 1) - (unsigned char)*(s2 - 1));
-	return (0);
+	if (!n)
+		return (0);
+	while (--n > 0 && (*s1 != '\0' || *s2 != '\0'))
+	{	
+		if (*s1 != *s2)
+			break ;
+		s1++;
+		s2++;
+	}
+	return ((unsigned char)*s1 - (unsigned char)*s2);
 }