+++ src/builtins/env.c
@@ -6,13 +6,13 @@
 /*   By: emadriga <emadriga@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/31 19:28:26 by emadriga          #+#    #+#             */
-/*   Updated: 2021/11/07 13:41:03 by emadriga         ###   ########.fr       */
+/*   Updated: 2021/11/09 12:58:44 by emadriga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
-#define ENV_NOT_OPTIONS_OR_ARGUMENT_ALLOWED \
- "Minishell's subject 'can't 'env with no options or arguments'\n"
+#define ENV_NOT_OPTIONS_OR_ARGUMENT_ALLOWED "Minishell's subject \
+'env with no options or arguments'\n"
 
 /**
  * * Get a default env
@@ -25,11 +25,11 @@ static void	get_default_env(t_str **env_list)
 
 	pwd_command = NULL;
 	getcwd(pwd, 100);
-	ft_lst_str_add_sorted(env_list, ft_strjoin(LITERAL_PWD_LIKE, pwd));
-	ft_lst_str_add_sorted(env_list, ft_strdup(LITERAL_SHLVL_LIKE_ONE));
-	ft_lst_str_add_sorted(env_list, ft_strdup(LITERAL_OLDPWD));
-	pwd_command = ft_strjoin(pwd, LITERAL_EXEC_MINISHELL);
-	ft_lst_str_add_sorted(env_list, ft_strjoin(LITERAL_LAST_CMD, pwd_command));
+	ft_lst_str_add_sorted(env_list, ft_strjoin(LIT_PWD_LIKE, pwd));
+	ft_lst_str_add_sorted(env_list, ft_strdup(LIT_SHLVL_LIKE_ONE));
+	ft_lst_str_add_sorted(env_list, ft_strdup(LIT_OLDPWD));
+	pwd_command = ft_strjoin(pwd, LIT_EXEC_MINISHELL);
+	ft_lst_str_add_sorted(env_list, ft_strjoin(LIT_LAST_CMD, pwd_command));
 	free(pwd_command);
 }
 
@@ -89,17 +89,17 @@ void	init_ms_env(char **env_vector, t_str **env_list)
 		shlvl = NULL;
 		while (env_vector[i] != NULL)
 		{
-			if (!ft_strncmp(env_vector[i], LITERAL_SHLVL_LIKE, 6))
+			if (!ft_strncmp(env_vector[i], LIT_SHLVL_LIKE, 6))
 				shlvl = ft_itoa(get_shlvl(ft_strchr(env_vector[i], '=') + 1));
 			else
 				ft_lst_str_add_sorted(env_list, ft_strdup(env_vector[i]));
 			i++;
 		}
 		if (!shlvl)
-			ft_lst_str_add_sorted(env_list, ft_strdup(LITERAL_SHLVL_LIKE_ONE));
+			ft_lst_str_add_sorted(env_list, ft_strdup(LIT_SHLVL_LIKE_ONE));
 		else
 		{
-			ft_lst_str_add_sorted(env_list, ft_strjoin(LITERAL_SHLVL_LIKE, shlvl));
+			ft_lst_str_add_sorted(env_list, ft_strjoin(LIT_SHLVL_LIKE, shlvl));
 			free(shlvl);
 		}
 	}